+++ packages/pluggableWidgets/html-element-web/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   We updated the light and dark icons and tiles for the widget.
 
+-   We changed colors in the structure mode preview for dark and light modes.
+
 ## [1.0.1] - 2023-01-05
 
 ### Fixed